Cast zstd error value to match printf format specifier used
There's no single printf format specifier that would match
the type in all environments. "%zu" would be the one with
compilers that support it, but not all compilers we use do
(notably it does not work on msys2)

See osdn #43355

+++ utility/ioz.c
@@ -1218,10 +1218,10 @@ const char *fz_strerror(fz_FILE *fp)
 
       if (NULL != cleartext) {
         fc_snprintf(zstderror, sizeof(zstderror), "ZSTD: \"%s\" (%ld)",
-                    cleartext, fp->u.zstd.error);
+                    cleartext, (long)fp->u.zstd.error);
       } else {
         fc_snprintf(zstderror, sizeof(zstderror), "ZSTD error %ld",
-                    fp->u.zstd.error);
+                    (long)fp->u.zstd.error);
       }
       return zstderror;
     }
